Fix exception when slave API does not match queue
There was an exception occurring in the slave monitor dashboard when
the CR /slave endpoint claimed that a slave was working on a certain
build, but the /queue endpoint did not contain info for that build.

This is an error state in ClusterRunner but it should not completely
kill the dashboard.